change frame output to vector

+++ lib/ofdm_adaptive_frame_bb_impl.h
@@ -63,9 +63,9 @@ private:
     constellation_type_t d_constellation;
     unsigned char d_fec_scheme;
     uint64_t d_tag_offset;
-    size_t d_frame_len;
+    int d_frame_len;
     pmt::pmt_t d_packet_len_tag;
-    size_t d_payload_carriers;
+    int d_payload_carriers;
     size_t d_bytes;
     unsigned char d_bps;
     bool d_waiting_full_frame;
